rename the file name in the file structure
