+++ pages/index.js
@@ -3,7 +3,6 @@ import BookList from "../components/BooksList/index.js";
 export default function HomePage() {
   return (
     <div>
-      <h1></h1>
       <BookList />
     </div>
   );
